make router responsible for prefetching
