+++ packages/kit/src/runtime/app/navigation.js
@@ -27,7 +27,7 @@ async function goto_(href, opts) {
 
 /** @param {string} href */
 function prefetch_(href) {
-	return renderer.prefetch(new URL(href, get_base_uri(document)));
+	return router.prefetch(new URL(href, get_base_uri(document)));
 }
 
 /** @param {string[]} [pathnames] */
