+++ packages/kit/src/runtime/client/utils.js
@@ -1,12 +1,3 @@
-/**
- * @param {Node} node
- * @returns {HTMLAnchorElement | SVGAElement}
- */
-export function find_anchor(node) {
-	while (node && node.nodeName.toUpperCase() !== 'A') node = node.parentNode; // SVG <a> elements have a lowercase name
-	return /** @type {HTMLAnchorElement | SVGAElement} */ (node);
-}
-
 /** @param {HTMLDocument} doc */
 export function get_base_uri(doc) {
 	let baseURI = doc.baseURI;
